Move the EventEncryptionAlgorithm into a more logical place
