+++ crates/matrix-sdk-crypto/src/olm/account.rs
@@ -61,9 +61,8 @@ use crate::{
                 EncryptedToDeviceEvent, OlmV1Curve25519AesSha2Content,
                 OlmV2Curve25519AesSha2Content, ToDeviceEncryptedEventContent,
             },
-            EventEncryptionAlgorithm,
         },
-        CrossSigningKey, DeviceKeys, OneTimeKey, SignedKey,
+        CrossSigningKey, DeviceKeys, EventEncryptionAlgorithm, OneTimeKey, SignedKey,
     },
     utilities::encode,
     CryptoStoreError, OlmError, SignatureError,
