+++ crates/matrix-sdk-crypto/src/olm/group_sessions/mod.rs
@@ -29,11 +29,8 @@ use vodozemac::{megolm::SessionKeyDecodeError, Curve25519PublicKey};
 
 use crate::types::{
     deserialize_curve_key,
-    events::{
-        forwarded_room_key::{ForwardedMegolmV1AesSha2Content, ForwardedRoomKeyContent},
-        EventEncryptionAlgorithm,
-    },
-    serialize_curve_key, SigningKey, SigningKeys,
+    events::forwarded_room_key::{ForwardedMegolmV1AesSha2Content, ForwardedRoomKeyContent},
+    serialize_curve_key, EventEncryptionAlgorithm, SigningKey, SigningKeys,
 };
 
 /// An error type for the creation of group sessions.
