+++ crates/matrix-sdk-crypto/src/types/events/forwarded_room_key.rs
@@ -21,10 +21,10 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use vodozemac::{megolm::ExportedSessionKey, Curve25519PublicKey, Ed25519PublicKey};
 
-use super::{EventEncryptionAlgorithm, EventType, ToDeviceEvent};
+use super::{EventType, ToDeviceEvent};
 use crate::types::{
     deserialize_curve_key, deserialize_curve_key_vec, deserialize_ed25519_key, serialize_curve_key,
-    serialize_curve_key_vec, serialize_ed25519_key,
+    serialize_curve_key_vec, serialize_ed25519_key, EventEncryptionAlgorithm,
 };
 
 /// The `m.forwarded_room_key` to-device event.
